+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.14.2] - 2024-01-29
+
 ### Changed
 
 - Cache hosted zone ID on creation
@@ -183,7 +185,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - changed: `app.giantswarm.io` label group was changed to `application.giantswarm.io`
 
-[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.14.1...HEAD
+[Unreleased]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.14.2...HEAD
+[0.14.2]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.14.1...v0.14.2
 [0.14.1]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.14.0...v0.14.1
 [0.14.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/giantswarm/aws-resolver-rules-operator/compare/v0.12.0...v0.13.0